reorganize: move classes to libs directory
