+++ src/dct/ai/Commander.lua
@@ -10,9 +10,9 @@ local containers = require("libs.containers")
 local enum       = require("dct.enum")
 local dctutils   = require("dct.utils")
 local Mission    = require("dct.ai.Mission")
-local Stats      = require("dct.Stats")
+local Stats      = require("dct.libs.Stats")
 local Command    = require("dct.Command")
-local Logger     = require("dct.Logger").getByName("Commander")
+local Logger     = dct.Logger.getByName("Commander")
 
 local function heapsort_tgtlist(assetmgr, owner, filterlist)
 	local tgtlist = assetmgr:getTargets(owner, filterlist)
